add missing test for rules implementation
